GUAC-1132: Refactor much of SynchronizedGuacamoleTunnel into AbstractGuacamaoleTunnel. Rename to SimpleGuacamoleTunnel.
